refactor(cli): split main.py into commands/ directory (closes #48)
- Move each subcommand into its own module under
  `provero/cli/commands/` (init, run, validate, profile, history,
  contract, watch, export, import_cmd)
- main.py now only wires the Typer app, registers commands, and owns
  shared options (--quiet, --debug, etc); shrinks from ~1175 to ~264
  lines
- Behavior unchanged; existing CLI tests pass without modification
